WIP: Refactor of Profiles to React Query
Why?
Right now zustand is being used to manage the network state of profiles
We refresh profiles whenever a new group is added/new conversation/group is refreshed

Theres some potential issues:
Groups can get large
Conversation lists can get large
New and interesting ways of using profiles in the app may be added (group invites for example)
These require adding additional checks if we should update the profiles etc.
What happens when a request fails? What happens if members change?
As we move towards open source it can be helpful to devs to have more consumable hooks, rather than tracking down exactly where the data is coming from and why it's refreshing/cached/missing

With React Query we get better out of the box retries, and can use a batching and deduplicating library to manage this for us
The main goal of this is that incase anything gets missed, we can have a foolproof fallback to refresh just in case